mechanism to self-loop on bursts
the previous implementation processed a single MQTT Message per update cycle. if things are coming in fast, a backlog accumulates. this mechanism re-triggers update if a message was received.

the example app has been augmented with some key functions; hit '2' to set the framerate to 2, it will exacerbate the behaviour. it 's' to toggle between self-loop and "old school" updates, and use 'b' to send bursts. when running at fps=2 and self_loop = false, the queueing behavior is evident.