threads support (native mosquitto + threadChannel)
the ofxThreadedMQTT class implements the native mosquitto pthread support, and provides a thread-safe threadChannel interface to the messages.

The class hierarchy is preferred to a constructor/setup variable as these bool are opaque and not self-documenting. A class named ofxThreadedMQTT is self-documenting (and the possible feature of live-toggling between threaded and non-threaded is probably not useful as the 2 approaches are no so directly compatible).

The upcoming mosquitto 2.1 supports the threads on Windows (currently only POSIX).

An example is provided.